+++ Frames/AceMainFrame.lua
@@ -5,8 +5,6 @@ local RS = RS_ADDON ---@type RestockerAddon
 ---@field frame table|nil AceGUI Frame widget
 ---@field addItemEditBox table|nil AceGUI EditBox widget
 local aceMainFrameModule = RsModule.aceMainFrameModule ---@type RsAceMainFrameModule
-local bankModule = RsModule.bankModule ---@type RsBankModule
-local eventsModule = RsModule.eventsModule ---@type RsEventsModule
 local restockerModule = RsModule.restockerModule ---@type RsRestockerModule
 
 local AceGUI = LibStub("AceGUI-3.0")
@@ -55,13 +53,6 @@ local function installDropTarget(frame)
       addCursorItem()
     end
   end)
-end
-
-local function refreshAfterBankSensitiveChange()
-  RS:Update()
-  if bankModule.bankIsOpen then
-    eventsModule.OnBankOpen(true)
-  end
 end
 
 local function createProfileList()
@@ -217,16 +208,6 @@ function aceMainFrameModule:CreateListHeader(parent)
   buyLabel:SetWidth(70)
   row:AddChild(buyLabel)
 
-  local toBankLabel = AceGUI:Create("Label")
-  toBankLabel:SetText("To bank")
-  toBankLabel:SetWidth(90)
-  row:AddChild(toBankLabel)
-
-  local fromBankLabel = AceGUI:Create("Label")
-  fromBankLabel:SetText("From bank")
-  fromBankLabel:SetWidth(105)
-  row:AddChild(fromBankLabel)
-
   local reputationLabel = AceGUI:Create("Label")
   reputationLabel:SetText("Vendor rep")
   reputationLabel:SetWidth(130)
@@ -262,7 +243,7 @@ function aceMainFrameModule:CreateItemRow(parent, item)
     item.amount = tonumber(value) or 0
     widget:SetText(tostring(item.amount))
     AceGUI:ClearFocus()
-    refreshAfterBankSensitiveChange()
+    RS:Update()
   end)
   row:AddChild(amountBox)
 
@@ -275,26 +256,6 @@ function aceMainFrameModule:CreateItemRow(parent, item)
     RS:Update()
   end)
   row:AddChild(buyCheckBox)
-
-  local toBankCheckBox = AceGUI:Create("CheckBox")
-  toBankCheckBox:SetLabel("")
-  toBankCheckBox:SetWidth(90)
-  toBankCheckBox:SetValue(item.stashTobank and true or false)
-  toBankCheckBox:SetCallback("OnValueChanged", function(_widget, _event, value)
-    item.stashTobank = value
-    refreshAfterBankSensitiveChange()
-  end)
-  row:AddChild(toBankCheckBox)
-
-  local fromBankCheckBox = AceGUI:Create("CheckBox")
-  fromBankCheckBox:SetLabel("")
-  fromBankCheckBox:SetWidth(105)
-  fromBankCheckBox:SetValue(item.restockFromBank and true or false)
-  fromBankCheckBox:SetCallback("OnValueChanged", function(_widget, _event, value)
-    item.restockFromBank = value
-    refreshAfterBankSensitiveChange()
-  end)
-  row:AddChild(fromBankCheckBox)
 
   local reputationDropDown = AceGUI:Create("Dropdown")
   reputationDropDown:SetList(reputationList, reputationOrder)